chore: refactoring of the Backup Modal (split to subcomponents)
